+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMKVAW8XEXEN9GK7STY41018"
+        "01KMMWCWK4MS87F1XRAAP6HPRP"
     )
     r"""The unique identifier of the data source"""