+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMKVAW8WNZ426PFKNZY5MV5J"
+        "01KMMWCWK3Z43ZXJ79M3J2QHYK"
     )
     r"""The unique identifier of the data source"""