+++ packages/orq-rc/RELEASES.md
@@ -6619,3 +6619,13 @@ Based on:
 - [python v4.7.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.21 - packages/orq-rc
+
+## 2026-03-26 10:55:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.22 - packages/orq-rc